+++ zappa/cli.py
@@ -78,7 +78,7 @@ class ZappaCLI(object):
     # CLI
     vargs = None
     command = None
-    command_env = None
+    stage_env = None
 
     # Zappa settings
     zappa = None
@@ -109,6 +109,7 @@ class ZappaCLI(object):
     exception_handler = None
     environment_variables = None
     authorizer = None
+    aws_kms_key_arn = ''
 
     stage_name_env_pattern = re.compile('^[a-zA-Z0-9_]+$')
 
@@ -133,7 +134,7 @@ class ZappaCLI(object):
             try:
                 stage_settings = dict(self.zappa_settings[stage].copy())
             except KeyError:
-                raise ClickException("Cannot extend settings for undefined environment '" + stage + "'.")
+                raise ClickException("Cannot extend settings for undefined stage '" + stage + "'.")
 
             extends_stage = self.zappa_settings[stage].get('extends', None)
             if not extends_stage:
@@ -190,9 +191,9 @@ class ZappaCLI(object):
         env_parser = argparse.ArgumentParser(add_help=False)
         me_group = env_parser.add_mutually_exclusive_group()
         all_help = ('Execute this command for all of our defined '
-                    'Zappa environments.')
+                    'Zappa stages.')
         me_group.add_argument('--all', action='store_true', help=all_help)
-        me_group.add_argument('command_env', nargs='?')
+        me_group.add_argument('stage_env', nargs='?')
 
         group = env_parser.add_argument_group()
         group.add_argument(
@@ -401,17 +402,17 @@ class ZappaCLI(object):
 
         # Parse the input
         # NOTE(rmoe): Special case for manage command
-        # The manage command can't have both command_env and command_rest
+        # The manage command can't have both stage_env and command_rest
         # arguments. Since they are both positional arguments argparse can't
         # differentiate the two. This causes problems when used with --all.
         # (e.g. "manage --all showmigrations admin" argparse thinks --all has
-        # been specified AND that command_env='showmigrations')
+        # been specified AND that stage_env='showmigrations')
         # By having command_rest collect everything but --all we can split it
         # apart here instead of relying on argparse.
         if args.command == 'manage' and not self.vargs.get('all'):
-            self.command_env = self.vargs['command_rest'].pop(0)
+            self.stage_env = self.vargs['command_rest'].pop(0)
         else:
-            self.command_env = self.vargs.get('command_env')
+            self.stage_env = self.vargs.get('stage_env')
 
         self.command = args.command
 
@@ -433,51 +434,52 @@ class ZappaCLI(object):
         # Load and Validate Settings File
         self.load_settings_file(self.vargs.get('settings_file'))
 
-        # Should we execute this for all environments, or just one?
-        all_environments = self.vargs.get('all')
-        environments = []
+        # Should we execute this for all stages, or just one?
+        all_stages = self.vargs.get('all')
+        stages = []
 
-        if all_environments: # All envs!
-            environments = self.zappa_settings.keys()
+        if all_stages: # All stages!
+            stages = self.zappa_settings.keys()
         else: # Just one env.
-            if not self.command_env:
-                # If there's only one environment defined in the settings,
+            if not self.stage_env:
+                # If there's only one stage defined in the settings,
                 # use that as the default.
                 if len(self.zappa_settings.keys()) == 1:
-                    environments.append(list(self.zappa_settings.keys())[0])
+                    stages.append(list(self.zappa_settings.keys())[0])
                 else:
-                    parser.error("Please supply an environment to interact with.")
+                    parser.error("Please supply an stage to interact with.")
             else:
-                environments.append(self.command_env)
+                stages.append(self.stage_env)
 
-        for environment in environments:
+        for stage in stages:
             try:
-                self.dispatch_command(self.command, environment)
+                self.dispatch_command(self.command, stage)
             except ClickException as e:
                 # Discussion on exit codes: https://github.com/Miserlou/Zappa/issues/407
                 e.show()
                 sys.exit(e.exit_code)
 
-    def dispatch_command(self, command, environment):
+    def dispatch_command(self, command, stage):
         """
-        Given a command to execute and environment,
+        Given a command to execute and stage,
         execute that command.
         """
 
-        self.api_stage = environment
+        self.api_stage = stage
 
         if command not in ['status', 'manage']:
             if not self.vargs['json']:
-                click.echo("Calling " + click.style(command, fg="green", bold=True) + " for environment " +
+                click.echo("Calling " + click.style(command, fg="green", bold=True) + " for stage " +
                            click.style(self.api_stage, bold=True) + ".." )
 
         # Explicity define the app function.
-        if self.vargs['app_function'] is not None:
+        # Related: https://github.com/Miserlou/Zappa/issues/832
+        if self.vargs.get('app_function', None):
             self.app_function = self.vargs['app_function']
 
         # Load our settings, based on api_stage.
         try:
-            self.load_settings(self.vargs['settings_file'])
+            self.load_settings(self.vargs.get('settings_file'))
         except ValueError as e:
             print("Error: {}".format(e.message))
             sys.exit(-1)
@@ -690,7 +692,9 @@ class ZappaCLI(object):
                 dead_letter_config=self.dead_letter_config,
                 timeout=self.timeout_seconds,
                 memory_size=self.memory_size,
-                runtime=self.runtime
+                runtime=self.runtime,
+                environment_variables=self.environment_variables,
+                aws_kms_key_arn=self.aws_kms_key_arn
             )
 
         # Schedule events for this deployment
@@ -833,7 +837,9 @@ class ZappaCLI(object):
                                                         vpc_config=self.vpc_config,
                                                         timeout=self.timeout_seconds,
                                                         memory_size=self.memory_size,
-                                                        runtime=self.runtime
+                                                        runtime=self.runtime,
+                                                        environment_variables=self.environment_variables,
+                                                        aws_kms_key_arn=self.aws_kms_key_arn
                                                     )
 
         # Finally, delete the local copy our zip package
@@ -1285,7 +1291,7 @@ class ZappaCLI(object):
 
         # Create Env
         while True:
-            click.echo("Your Zappa configuration can support multiple production environments, like '" +
+            click.echo("Your Zappa configuration can support multiple production stages, like '" +
                        click.style("dev", bold=True)  + "', '" + click.style("staging", bold=True)  + "', and '" +
                        click.style("production", bold=True)  + "'.")
             env = input("What do you want to call this environment (default 'dev'): ") or "dev"
@@ -1293,7 +1299,7 @@ class ZappaCLI(object):
                 self.check_stage_name(env)
                 break
             except ValueError:
-                click.echo(click.style("Environment names must match a-zA-Z0-9_", fg="red"))
+                click.echo(click.style("Stage names must match a-zA-Z0-9_", fg="red"))
 
         # Detect AWS profiles and regions
         # If anyone knows a more straightforward way to easily detect and parse AWS profiles I'm happy to change this, feels like a hack
@@ -1426,7 +1432,8 @@ class ZappaCLI(object):
 
             env_bucket = bucket
             if global_deployment:
-                env_bucket = bucket.replace('-', '_') + '_' + env_name
+            # `zappa init` doesn't generate compatible s3_bucket names #828
+                env_bucket = (bucket + '-' + env_name).replace('_', '-')
 
             env_zappa_settings = {
                 env_name: {
@@ -1567,7 +1574,7 @@ class ZappaCLI(object):
 
         # Let's Encrypt
         if not cert_location and not cert_arn:
-            from letsencrypt import get_cert_and_update_domain, cleanup
+            from zappa.letsencrypt import get_cert_and_update_domain, cleanup
             cert_success = get_cert_and_update_domain(
                     self.zappa,
                     self.lambda_name,
@@ -1717,16 +1724,16 @@ class ZappaCLI(object):
         # Load up file
         self.load_settings_file(settings_file)
 
-        # Make sure that the environments are valid names:
+        # Make sure that the stages are valid names:
         for stage_name in self.zappa_settings.keys():
             try:
                 self.check_stage_name(stage_name)
             except ValueError:
                 raise ValueError("API stage names must match a-zA-Z0-9_ ; '{0!s}' does not.".format(stage_name))
 
-        # Make sure that this environment is our settings
+        # Make sure that this stage is our settings
         if self.api_stage not in self.zappa_settings.keys():
-            raise ClickException("Please define '{0!s}' in your Zappa settings.".format(self.api_stage))
+            raise ClickException("Please define stage '{0!s}' in your Zappa settings.".format(self.api_stage))
 
         # We need a working title for this project. Use one if supplied, else cwd dirname.
         if 'project_name' in self.stage_config: # pragma: no cover
@@ -1742,7 +1749,7 @@ class ZappaCLI(object):
         #           And various others from Slack.
         self.lambda_name = slugify.slugify(self.project_name + '-' + self.api_stage)
 
-        # Load environment-specific settings
+        # Load stage-specific settings
         self.s3_bucket_name = self.stage_config.get('s3_bucket', "zappa-" + ''.join(random.choice(string.ascii_lowercase + string.digits) for _ in range(9)))
         self.vpc_config = self.stage_config.get('vpc_config', {})
         self.memory_size = self.stage_config.get('memory_size', 512)
@@ -1784,6 +1791,7 @@ class ZappaCLI(object):
         self.check_environment(self.environment_variables)
         self.authorizer = self.stage_config.get('authorizer', {})
         self.runtime = self.stage_config.get('runtime', get_runtime_from_python_version())
+        self.aws_kms_key_arn = self.stage_config.get('aws_kms_key_arn', '')
 
         desired_role_name = self.lambda_name + "-ZappaLambdaExecutionRole"
         self.zappa = Zappa( boto_session=session,
@@ -2002,7 +2010,7 @@ class ZappaCLI(object):
                 env_dict['AWS_REGION'] = self.aws_region
             env_dict.update(dict(self.environment_variables))
 
-            # Environement variable keys can't be Unicode
+            # Environment variable keys can't be Unicode
             # https://github.com/Miserlou/Zappa/issues/604
             try:
                 env_dict = dict((k.encode('ascii'), v) for (k, v) in env_dict.items())