+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 # Zappa Changelog
 
+## 0.42.0
+* Cached manylinux wheel installed
+* New dependency installation formatting
+* Clarify "stage" vs "environment" terminology in code
+* Fix problem with capitalized packages
+* Delete local package if using wheels version. This saves several MBs on package size in some cases (e.g. numpy).
+* Thanks to @mcrowson, @nikbora and @schuyler1d
+
+## 0.41.3
+* Various Python3 fixes
+* Remove some dead code
+* More binary package fixes thanks to and @nikbora and @bxm156
+* Improved async tasks thanks to @schuyler1d
+* Various small changes
+
+## 0.41.2
+* Support for new `lambda-packages` format (Python3 support)
+* Fix `setup.py` on Windows Python3
+* Fix #818 - python3 import for LE
+* Support AWS-specific environment variables (with KMS)
+
 ## 0.41.1
 * Add `template` command
 * Add `--json` in more places